+++ src/main/java/toy/bookchat/bookchat/BookChatApplication.java
@@ -4,8 +4,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import toy.bookchat.bookchat.config.aws.StorageProperties;
-import toy.bookchat.bookchat.config.security.JwtTokenProperties;
-import toy.bookchat.bookchat.config.security.OAuth2Properties;
+import toy.bookchat.bookchat.config.token.JwtTokenProperties;
+import toy.bookchat.bookchat.config.token.OAuth2Properties;
 
 @SpringBootApplication
 @EnableConfigurationProperties({JwtTokenProperties.class, OAuth2Properties.class,
